moved admin views to its own file
